+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 ## 24.7.4
 * Mitigated an issue with the feedback widget that caused it to display a blank white page on devices running iOS versions earlier than 17
 
+* Added `CountlyFeedbacks:` interface with new view methods (Access with `Countly.sharedInstance.feedback`):
+    * Method to present feedback widget (wih an optional widget selector(name, ID or tag) string and a Callback):
+        * `presentNPS`
+        * `presentSurvey`
+        * `presentRating`
+    * `getAvailableFeedbackWidgets` method to retrieve available feedback widgets with a completion handler.
+    
+* Deprecated `getFeedbackWidgets` method, you should use `[feedback getAvailableFeedbackWidgets:]` method instead
+
 ## 24.7.3
 * Added current view names to event segmentation based on the `enablePreviousNameRecording` (Experimental!)
 * Updated the SDK to ensure compatibility with the latest server response models